fix: JavaScript Uncaught ReferenceError: resizeIframe
Uncaught ReferenceError: resizeIframe is not defined on line 102 of your index.html file.

This is happening because there's an iframe with an onload="resizeIframe(this)" attribute, but the resizeIframe function isn't defined anywhere in your code.

The function we added will automatically resize the iframe to fit its content, which should prevent scrollbars and make the embedded map display properly.

The resizeIframe function works by:
- Taking the iframe element as an input parameter
- Accessing the iframe's content window
- Getting the scroll height of the document inside the iframe
- Setting the iframe's height to match that scroll height

This is a common pattern for making iframes responsive to their content.

The resizeIframe function is being used on line 112 of your index.html file. It's attached to an iframe that loads a US map from "maps/us_map.html".